Completely reworked lasers + added a laser reflector.
Remade the laser system from the ground up and the emitter is the only object that creates the laser. Each laser adjuster (reflect, refract) now only acts as a reference object for the laser emitter to adjust its laser direction.

Previously, each laser adjust object could only handle the laser hitting it once otherwise it would cause an overflow error. The laser can now bounce off these objects an infinite amount of times.

This system is now easier to create and work with and is much more efficient.

Also added a reflect object which will act like a mirror for the laser.